Remove widget links
I removed the URLs from the links of the widget and I replaced them with
a # to avoid page reload when the widget are clicked. I am not use yet
if there's a way to keep the widget but without being clickable. I will
take a look into that in the future.